Consolidate Google Auto dependency objects
