+++ buildSrc/src/main/kotlin/io/spine/internal/dependency/Auto.kt
@@ -27,7 +27,21 @@
 package io.spine.internal.dependency
 
 // https://github.com/google/auto
+object AutoCommon {
+    private const val version = "1.2.1"
+    const val lib = "com.google.auto:auto-common:${version}"
+}
+
+// https://github.com/google/auto
+object AutoService {
+    private const val version = "1.0.1"
+    const val annotations = "com.google.auto.service:auto-service-annotations:${version}"
+    @Suppress("unused")
+    const val processor   = "com.google.auto.service:auto-service:${version}"
+}
+
+// https://github.com/google/auto
 object AutoValue {
-    private const val version = "1.9"
+    private const val version = "1.10.1"
     const val annotations = "com.google.auto.value:auto-value-annotations:${version}"
 }
